% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1Cellbio.R
\name{as.Seurat.1CB.1CellbioData}
\alias{as.Seurat.1CB.1CellbioData}
\title{S3 method for converting 1CellbioData to Seurat object}
\usage{
\method{as.Seurat}{`1CB.1CellbioData`}(object, ...)
}
\arguments{
\item{object}{A 1CellbioData object}

\item{...}{Additional arguments (not used)}

\item{rownames}{Character string specifying the column name in row data to use as gene names (required)}

\item{colnames}{Character string specifying the column name in column data to use as cell names (required)}
}
\value{
A Seurat object
}
\description{
S3 method for converting 1CellbioData to Seurat object
}
