% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1Cellbio.R
\name{as.Seurat}
\alias{as.Seurat}
\title{Convert 1CellbioData to Seurat object}
\usage{
as.Seurat(object, ...)
}
\arguments{
\item{object}{A 1CellbioData object}

\item{...}{Additional arguments (not used)}
}
\value{
A Seurat object
}
\description{
Convert 1CellbioData to Seurat object
}
