% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{as.Seurat.StereoData}
\alias{as.Seurat.StereoData}
\title{Convert StereoData to Seurat object}
\usage{
\method{as.Seurat}{StereoData}(object, ...)
}
\arguments{
\item{object}{A StereoData object}

\item{...}{Additional arguments}
}
\value{
A Seurat object
}
\description{
Convert StereoData to Seurat object
}
