% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1Cellbio.R
\name{as.SingleCellExperiment.1CellbioData}
\alias{as.SingleCellExperiment.1CellbioData}
\title{Convert 1CellbioData to SingleCellExperiment object}
\usage{
\method{as.SingleCellExperiment}{`1CellbioData`}(object, ...)
}
\arguments{
\item{object}{A 1CellbioData object}

\item{...}{Additional arguments (not used)}
}
\value{
A SingleCellExperiment object
}
\description{
Convert 1CellbioData to SingleCellExperiment object
}
