% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1Cellbio.R
\name{as.SingleCellExperiment}
\alias{as.SingleCellExperiment}
\title{Convert 1CellbioData to SingleCellExperiment object}
\usage{
as.SingleCellExperiment(object, ...)
}
\arguments{
\item{object}{A 1CellbioData object}

\item{...}{Additional arguments (not used)}
}
\value{
A SingleCellExperiment object
}
\description{
Convert 1CellbioData to SingleCellExperiment object
}
