% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{as.SingleCellExperiment.StereoData}
\alias{as.SingleCellExperiment.StereoData}
\title{Convert StereoData to SingleCellExperiment object}
\usage{
\method{as.SingleCellExperiment}{StereoData}(object, ...)
}
\arguments{
\item{object}{A StereoData object}

\item{...}{Additional arguments}
}
\value{
A SingleCellExperiment object
}
\description{
Convert StereoData to SingleCellExperiment object
}
