% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{build_lookup_tables}
\alias{build_lookup_tables}
\title{Build lookup tables for efficient gene ID searching}
\usage{
build_lookup_tables(mappings)
}
\arguments{
\item{mappings}{List of mapping data.tables}
}
\value{
List of lookup tables (gene_id -> gesel_id)
}
\description{
Build lookup tables for efficient gene ID searching
}
\keyword{internal}
