% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_sparse_matrix}
\alias{construct_sparse_matrix}
\title{Convert sparse matrix from HDF5 to dgCMatrix}
\usage{
construct_sparse_matrix(data, indices, indptr, shape)
}
\arguments{
\item{data}{Vector of non-zero values}

\item{indices}{Vector of column indices for each non-zero value}

\item{indptr}{Vector of index pointers for each row}

\item{shape}{Vector of matrix dimensions (rows, columns)}
}
\value{
A dgCMatrix object
}
\description{
This function converts a sparse matrix stored in HDF5 format (10x format) to a dgCMatrix
}
