% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{convert_gene_lists_to_gesel}
\alias{convert_gene_lists_to_gesel}
\title{Convert gene lists to gesel IDs}
\usage{
convert_gene_lists_to_gesel(gene_lists, mappings, species = NULL)
}
\arguments{
\item{gene_lists}{List of gene lists}

\item{mappings}{List of mapping data.tables}

\item{species}{Species identifier for loading pre-built lookup tables}
}
\value{
List of converted gene lists with gesel IDs
}
\description{
Convert gene lists to gesel IDs
}
\keyword{internal}
