% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x_mtx.R
\name{create_h5ad_from_integrated_data}
\alias{create_h5ad_from_integrated_data}
\title{Create h5ad file from integrated data}
\usage{
create_h5ad_from_integrated_data(
  matrix,
  cell_metadata,
  gene_metadata,
  output_file,
  verbose
)
}
\arguments{
\item{matrix}{Integrated sparse matrix (genes x cells)}

\item{cell_metadata}{Cell metadata data.frame}

\item{gene_metadata}{Gene metadata data.frame}

\item{output_file}{Output h5ad file path}

\item{verbose}{Whether to print verbose messages}
}
\description{
Create h5ad file from integrated data
}
\keyword{internal}
