% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{download_gene_mapping}
\alias{download_gene_mapping}
\title{Download gene mapping file}
\usage{
download_gene_mapping(species, type)
}
\arguments{
\item{species}{Species identifier}

\item{type}{Type of mapping (symbol, entrez, ensembl)}
}
\value{
Data.table with downloaded mapping data
}
\description{
Download gene mapping file
}
\keyword{internal}
