% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{encode_delta_line}
\alias{encode_delta_line}
\title{差分编码函数}
\usage{
encode_delta_line(arr)
}
\arguments{
\item{arr}{Array of integers to encode}
}
\value{
Delta-encoded string
}
\description{
差分编码函数
}
\keyword{internal}
