% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{generate_collections_file}
\alias{generate_collections_file}
\title{生成collections文件}
\usage{
generate_collections_file(name, desc, species, size, output_dir)
}
\arguments{
\item{name}{Collection name}

\item{desc}{Collection description}

\item{species}{Species identifier}

\item{size}{Number of gene sets}

\item{output_dir}{Output directory}
}
\description{
生成collections文件
}
\keyword{internal}
