% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{generate_gene2set_file}
\alias{generate_gene2set_file}
\title{生成gene2set文件}
\usage{
generate_gene2set_file(converted_lists, species, output_dir)
}
\arguments{
\item{converted_lists}{List of converted gene lists}

\item{species}{Species identifier}

\item{output_dir}{Output directory}
}
\description{
生成gene2set文件
}
\keyword{internal}
