% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{generate_gesel_files}
\alias{generate_gesel_files}
\title{生成gesel格式的完整文件集}
\usage{
generate_gesel_files(
  converted_lists,
  set_names,
  set_descriptions,
  species,
  collection_name,
  collection_desc,
  output_dir
)
}
\arguments{
\item{converted_lists}{List of converted gene lists}

\item{set_names}{Vector of gene set names}

\item{set_descriptions}{Vector of gene set descriptions}

\item{species}{Species identifier}

\item{collection_name}{Name of the collection}

\item{collection_desc}{Description of the collection}

\item{output_dir}{Output directory}
}
\description{
生成gesel格式的完整文件集
}
\keyword{internal}
