% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{generate_set2gene_file}
\alias{generate_set2gene_file}
\title{生成set2gene文件（差分编码）}
\usage{
generate_set2gene_file(converted_lists, species, output_dir)
}
\arguments{
\item{converted_lists}{List of converted gene lists}

\item{species}{Species identifier}

\item{output_dir}{Output directory}
}
\description{
生成set2gene文件（差分编码）
}
\keyword{internal}
