% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{generate_sets_file}
\alias{generate_sets_file}
\title{生成sets文件}
\usage{
generate_sets_file(names, descriptions, sizes, species, output_dir)
}
\arguments{
\item{names}{Gene set names}

\item{descriptions}{Gene set descriptions}

\item{sizes}{Gene set sizes}

\item{species}{Species identifier}

\item{output_dir}{Output directory}
}
\description{
生成sets文件
}
\keyword{internal}
