% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup_tables.R
\name{get_species_lookup_table}
\alias{get_species_lookup_table}
\title{Get lookup table for specific species}
\usage{
get_species_lookup_table(species, rdata_file = "data/gene_lookup_tables.RData")
}
\arguments{
\item{species}{Species identifier}

\item{rdata_file}{Path to the RData file}
}
\value{
Lookup tables for the species
}
\description{
Get lookup table for specific species
}
