% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icb_anndata_to_csparse}
\alias{icb_anndata_to_csparse}
\title{Robustly obtain CsparseMatrix from AnnData matrix/layer}
\usage{
icb_anndata_to_csparse(obj, transpose = FALSE, verbose = FALSE)
}
\arguments{
\item{obj}{Anndata matrix-like (possibly Python object)}

\item{transpose}{Whether to transpose after conversion}

\item{verbose}{Whether to print diagnostic}
}
\description{
Robustly obtain CsparseMatrix from AnnData matrix/layer
}
\keyword{internal}
