% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icb_is_py_sparse}
\alias{icb_is_py_sparse}
\title{Check if a Python object is scipy.sparse matrix}
\usage{
icb_is_py_sparse(py_obj)
}
\description{
Check if a Python object is scipy.sparse matrix
}
\keyword{internal}
