% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icb_make_unique}
\alias{icb_make_unique}
\title{Make names unique or error}
\usage{
icb_make_unique(names, strategy = c("make_unique", "error"), sep = "-")
}
\arguments{
\item{names}{Character vector}

\item{strategy}{One of "make_unique" (default) or "error"}

\item{sep}{Suffix separator when making unique}
}
\description{
Make names unique or error
}
\keyword{internal}
