% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{icb_py_sparse_to_dgc}
\alias{icb_py_sparse_to_dgc}
\title{Convert scipy.sparse (CSR/CSC) to dgCMatrix efficiently}
\usage{
icb_py_sparse_to_dgc(py_obj)
}
\description{
Convert scipy.sparse (CSR/CSC) to dgCMatrix efficiently
}
\keyword{internal}
