% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x_mtx.R
\name{integrate_samples}
\alias{integrate_samples}
\title{Integrate matrices from multiple samples}
\usage{
integrate_samples(matrices, cell_metadata_list, gene_names, verbose)
}
\arguments{
\item{matrices}{List of sparse matrices (one per sample)}

\item{cell_metadata_list}{List of cell metadata data.frames}

\item{gene_names}{Vector of gene names to use for integration}

\item{verbose}{Whether to print verbose messages}
}
\value{
A list containing integrated matrix and metadata
}
\description{
Integrate matrices from multiple samples
}
\keyword{internal}
