% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{load_custom_mapping_file}
\alias{load_custom_mapping_file}
\title{Load custom mapping file}
\usage{
load_custom_mapping_file(file_path, type)
}
\arguments{
\item{file_path}{Path to the mapping file}

\item{type}{Type of mapping (symbol, entrez, ensembl)}
}
\value{
Data.table with mapping data
}
\description{
Load custom mapping file
}
\keyword{internal}
