% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lookup_tables.R
\name{load_gene_lookup_tables}
\alias{load_gene_lookup_tables}
\title{Load pre-built lookup tables}
\usage{
load_gene_lookup_tables(rdata_file = "data/gene_lookup_tables.RData")
}
\arguments{
\item{rdata_file}{Path to the RData file containing pre-built lookup tables}
}
\value{
List of lookup tables
}
\description{
Load pre-built lookup tables
}
