% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{load_local_mapping_file}
\alias{load_local_mapping_file}
\title{Load local mapping file from package data}
\usage{
load_local_mapping_file(file_path, type)
}
\arguments{
\item{file_path}{Path to the local mapping file}

\item{type}{Type of mapping (symbol, entrez, ensembl)}
}
\value{
Data.table with mapping data
}
\description{
Load local mapping file from package data
}
\keyword{internal}
