% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1Cellbio.R
\name{make_unique_names}
\alias{make_unique_names}
\title{Helper function to make unique names}
\usage{
make_unique_names(names)
}
\arguments{
\item{names}{Character vector of names}
}
\value{
Character vector with unique names
}
\description{
Helper function to make unique names
}
