% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{parse_gmt_data}
\alias{parse_gmt_data}
\title{Parse GMT data}
\usage{
parse_gmt_data(gmt_data)
}
\arguments{
\item{gmt_data}{Data.table containing GMT file contents}
}
\value{
List with set_names, set_descriptions, and gene_lists
}
\description{
Parse GMT data
}
\keyword{internal}
