% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{plot_cells_with_borders}
\alias{plot_cells_with_borders}
\title{Plot cells with borders}
\usage{
plot_cells_with_borders(
  object,
  color_by = NULL,
  show_borders = TRUE,
  border_color = "black",
  border_size = 0.1,
  point_size = 0.5
)
}
\arguments{
\item{object}{A StereoData object or Seurat object with cell borders}

\item{color_by}{Variable to color cells by (from cell metadata)}

\item{show_borders}{Whether to show cell borders}

\item{border_color}{Color for cell borders}

\item{border_size}{Line width for cell borders}

\item{point_size}{Size of cell center points}
}
\value{
A ggplot object
}
\description{
Plot cells with borders
}
