% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1Cellbio.R
\name{read1Cellbio}
\alias{read1Cellbio}
\title{Read 1Cellbio results from zip file}
\usage{
read1Cellbio(file_path)
}
\arguments{
\item{file_path}{Path to the zip file containing 1Cellbio results}
}
\value{
A 1CellbioData object
}
\description{
This function reads 1Cellbio pipeline results directly from a zip file
and creates a 1CellbioData object that can be converted to Seurat or SingleCellExperiment.
}
\examples{
\dontrun{
data <- read1Cellbio("path/to/results.zip")
sce <- as.SingleCellExperiment.1CB(data)
seurat <- as.Seurat.1CB(data)
}
}
