% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_10x_mtx.R
\name{read_csv_sample_info}
\alias{read_csv_sample_info}
\title{Read CSV file with sample information}
\usage{
read_csv_sample_info(csv_file)
}
\arguments{
\item{csv_file}{Path to the CSV file}
}
\value{
A data.frame with sample information
}
\description{
Read CSV file with sample information
}
\keyword{internal}
