% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{read_gef_bins}
\alias{read_gef_bins}
\title{Read GEF file with square bin data}
\usage{
read_gef_bins(h5file, bin_size, gene_list = NULL, region = NULL)
}
\description{
Read GEF file with square bin data
}
\keyword{internal}
