% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{read_gef_cell_borders}
\alias{read_gef_cell_borders}
\title{Read cell border information from GEF cellBin group}
\usage{
read_gef_cell_borders(cellbin_group, cell_info)
}
\description{
Read cell border information from GEF cellBin group
}
\keyword{internal}
