% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{read_gef_cell_info}
\alias{read_gef_cell_info}
\title{Read cell information from GEF cellBin group}
\usage{
read_gef_cell_info(cellbin_group, max_cells = NULL, region = NULL)
}
\description{
Read cell information from GEF cellBin group
}
\keyword{internal}
