% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{read_gef_cellbins}
\alias{read_gef_cellbins}
\title{Read GEF file with cell bin data}
\usage{
read_gef_cellbins(
  h5file,
  gene_list = NULL,
  region = NULL,
  max_cells = NULL,
  include_cellborder = TRUE
)
}
\description{
Read GEF file with cell bin data
}
\keyword{internal}
