% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{read_gef_expression_matrix}
\alias{read_gef_expression_matrix}
\title{Read expression matrix from GEF cellBin group (fast vectorized construction)}
\usage{
read_gef_expression_matrix(cellbin_group, cell_info, gene_info)
}
\description{
Read expression matrix from GEF cellBin group (fast vectorized construction)
}
\keyword{internal}
