% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{read_gef_gene_info}
\alias{read_gef_gene_info}
\title{Read gene information from GEF cellBin group}
\usage{
read_gef_gene_info(cellbin_group, gene_list = NULL)
}
\description{
Read gene information from GEF cellBin group
}
\keyword{internal}
