% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gef.R
\name{summary.StereoData}
\alias{summary.StereoData}
\title{Get summary information about StereoData object}
\usage{
summary.StereoData(object)
}
\arguments{
\item{object}{A StereoData object}
}
\description{
Get summary information about StereoData object
}
