% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_csparse}
\alias{to_csparse}
\title{Ensure a matrix is CsparseMatrix}
\usage{
to_csparse(x)
}
\arguments{
\item{x}{A matrix-like object (dense or sparse)}
}
\value{
CsparseMatrix
}
\description{
Ensure a matrix is CsparseMatrix
}
\keyword{internal}
