% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gmt.R
\name{validate_mapping_format}
\alias{validate_mapping_format}
\title{Validate mapping file format}
\usage{
validate_mapping_format(data, type)
}
\arguments{
\item{data}{Data.table to validate}

\item{type}{Type of mapping}
}
\description{
Validate mapping file format
}
\keyword{internal}
