% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1Cellbio.R
\name{as.SingleCellExperiment.1CellbioData}
\alias{as.SingleCellExperiment.1CellbioData}
\title{Convert 1CellbioData to SingleCellExperiment object (DEPRECATED)}
\usage{
\method{as.SingleCellExperiment}{`1CellbioData`}(
  object,
  rownames = NULL,
  colnames = NULL,
  ...,
  name_conflict = c("make_unique", "error")
)
}
\arguments{
\item{object}{A 1CellbioData object}

\item{rownames}{Character string specifying the column name in row data to use as gene names (required)}

\item{colnames}{Character string specifying the column name in column data to use as cell names (required)}

\item{...}{Additional arguments (not used)}
}
\value{
A SingleCellExperiment object
}
\description{
Convert 1CellbioData to SingleCellExperiment object (DEPRECATED)
}
\keyword{internal}
