% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_hdf5_dataframe}
\alias{read_hdf5_dataframe}
\title{Read HDF5 data frame}
\usage{
read_hdf5_dataframe(file_path)
}
\arguments{
\item{file_path}{Path to the HDF5 file}
}
\value{
A data.frame object
}
\description{
This function reads a data frame from HDF5 file
}
