% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read1Cellbio.R, R/utils.R
\name{read_hdf5_sparse_matrix}
\alias{read_hdf5_sparse_matrix}
\title{Read a sparse matrix from HDF5 file}
\usage{
read_hdf5_sparse_matrix(file_path)

read_hdf5_sparse_matrix(file_path)
}
\arguments{
\item{file_path}{Path to the HDF5 file}
}
\value{
A sparse matrix

A dgCMatrix object
}
\description{
This function reads a sparse matrix from HDF5 file and returns a dgCMatrix
}
